chore: split network files
