+++ internal/swarm/network_manager_test.go
